requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic>=2.5.3
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-text-splitters>=1.0.0
